Refactor: Make id and chatId nullable in ChatMessage and ChatMessageDto
This commit modifies the `ChatMessage` data class and its corresponding DTO, `ChatMessageDto`, to allow `id` and `chatId` properties to be nullable.

This change provides flexibility for scenarios where these identifiers might not be immediately available or applicable, such as when creating a new message before it's persisted and assigned an ID by the backend.